separation of app.js from index.js containing http server and socket
